requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
groq==0.11.0
jpholiday==1.0.3
dateparser==1.2.0
python-dotenv==1.0.0
